requirements.txt
openai-agents
tenacity
